api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pyzxing
Pillow
requests
python-multipart
openai>=1.0.0
python-jose[cryptography]
passlib[bcrypt]
email-validator